Split clusters.go into AWS/Azure/helpers files, fix docs
- clusters.go: step struct + Run dispatch (~84 lines)
- clusters_aws.go: all AWS-specific logic (~2617 lines)
- clusters_azure.go: all Azure-specific logic (~1152 lines)
- clusters_helpers.go: buildAWSClustersResourceTags, buildAzureRequiredTags,
  buildResourceTagsWithExtra (~59 lines)

Also:
- Fix comment: "per-release kubeconfigs" → "per-cluster kubeconfigs"
- Fix step-dependencies.md: clusters step is Go for both AWS and Azure
  (not "Azure still uses Python Pulumi for some resources")
